requirements.txt
pyaudio==0.2.11
numpy==1.24.3
webrtcvad==2.0.10
torch==2.0.1
torchaudio==2.0.2
transformers==4.36.0
openai-whisper==20231117
soundfile==0.12.1
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0
scipy==1.10.1
librosa==0.10.0
pygame==2.5.0
TTS==0.21.1